feat(bench): Add direct API benchmark to filter_parallel.rs (#4297)
* refactor(bench): Separate morsel_operation_benchmark.rs into per-operation files

Split the combined 35KB morsel_operation_benchmark.rs into dedicated files:
- filter_parallel.rs - Filter/WHERE benchmarks
- scan_parallel.rs - Parallel scan benchmarks
- groupby_parallel.rs - Group by benchmarks
- join_probe_parallel.rs - Hash join probe phase (build in hash_join_parallel.rs)
- aggregate_parallel.rs - Aggregate/reduce benchmarks

Benefits:
- Easier debugging - can run/modify one benchmark without affecting others
- Faster iteration - only recompile what you're working on
- Clear ownership - each file has a single purpose
- Consistency - all operations follow the same organizational pattern

Closes #4290



* feat(bench): Add direct API benchmark to filter_parallel.rs

Restores the bench_morsel_filter_direct() function that was lost when
morsel_operation_benchmark.rs was split into per-operation files in PR #4293.

Changes:
- Add bench_morsel_filter_direct() for testing morsel_parallel_filter API directly
- Add BENCHMARK_FILTER env var to run specific benchmarks (morsel_filter, sql_filter)
- Update documentation to explain the two benchmark types:
  - Direct API (row-oriented): Tests morsel work-stealing performance
  - SQL-based (columnar): Routes through SIMD columnar execution

This benchmark is important for measuring raw morsel parallelism without
the ~100x speedup from columnar SIMD execution.



---------
